flujo crear Categoria con tests

+++ ventacontable/microservicio/dominio/src/main/java/com/ventacontable/producto/puerto/repositorio/RepositorioCategoria.java
@@ -9,7 +9,7 @@ public interface RepositorioCategoria {
      * @param categoria
      * @return el id generado
      */
-    Long crear(Categoria categoria);
+    int crear(Categoria categoria);
 
     /**
      * Permite eliminar una categoria
